Acceleration Structure Rework Done
