+++ src/rendering/ressources/meson.build
@@ -1,0 +1,3 @@
+src += files(
+    'AccelerationStructure.cpp',
+)
